CL Scrum 1.4
This update allows you to change the names of the lists or add more lists to your board.
It has not been tested very well so do note that bugs are likely.

If you want to use V 1.4 first you need to run the update.exe in the directory of the CL Scrum.exe. This update program will delete files not used by the new version and create ones that are used by the newer version.